Add like functionality
This commit adds the Likes table that acts similar to a polimorphic
table but it's not actually connected to another AR table.

The article class holds an id and a likes column now. Before rendering
all articles we search for any Like records in db and pass them to the
article class. If we don't find any likes we default the likes column to
0

We create a like record when the user attepts to like an article. The
Like AR class has an increment counter method that is called in the
Likes controller.

We then use turbo frames to only update a specific part of the page and
not reload the whole page. Providing a better experience for the user.